worker/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
psycopg2-binary>=2.9.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
